basics/syntax: codify tab settings
Markdown.pl (incorrectly) expands hard tabs to spaces
using a tab stops set 4 spaces apart.

While it would be nice to fix this for code blocks, it
expands the tabs to spaces before doing anything else
so it would be non-trivial to do so.

Instead explain this deviant behavior in the help
files.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>