Markdown.pl: add support for --imageroot <prefix>
If --imageroot <prefix> is used, then <prefix> will be
prefixed to all generated URLs that are not absolute URLs
(i.e. any that do not start with a scheme or '//' and are
not email addresses) and end in an image extension.  This
will override any --htmlroot <prefix> setting in that case.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>