Markdown.pl: modularize for reusability
Options may now be passed in to the Markdown function and it
may be called multiple times with different options on different
texts with no unwanted interaction between the calls.

Simply `require "Markdown.pl"` and then call Markdown::Markdown.

Or something like this will import the `Markdown` function
regardless of whether it's available in Markdown.pl or Markdown.pm:

  BEGIN {eval {require "Markdown.pl"}}
  use Markdown qw(Markdown);

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>