Rename various files
Get rid of ".text" extension.
Use standard "README" and "LICENSE" names for those.
The help files are markdown so make them have a .md extension.

Markdown.pl is kept as is (rather than changing it to markdown.pl)
because it's also a Perl module (Markdown) and Perl module names
typically start with an uppercase letter.  Were it to be renamed
to a ".pm" it would end up being Markdown.pm NOT markdown.pm.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>