Markdown.pl: output deterministic documents
Obscuring email addresses is all very nice, but
outputting a different document every time for the
same input is not.

It screws up caching and last modified checks and
is a very bad thing to do.

Instead continue to obscure email addresses, but
arrange for the same obscurity to be used when the
same input file is processed repeatedly by Markdown.pl
on the same machine.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>