Markdown.pl: introduce --tabwidth=<num> option
The --tabwidth=<num> option only affects the width to which
tabs are expanded.  It does NOT affect the number of spaces
required to start a new indent level.  That remains set at 4
no matter what value is used for the --tabwidth=<num> option.

With this change it's now, finally, possible to have proper
tab expansion without breaking the "4 spaces per indent level"
rule.

Note that backticks-delimited code blocks will always expand
their tabs to 8-character tab stop positions no matter what
value is used for the --tabwidth=<num> option.

With this change the default expansion width for tabs when
Markdown.pl is run from the command line is now 8.

When used as a module the default is still 4, but that's
easily changed by passing in a suitable option.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>