Markdown.pl: recognize top-level lists better
The change in version 1.0.1 to fix "a bug where lines in the
middle of hard-wrapped paragraphs, which lines look like the
start of a list item, would accidentally trigger the creation
of a list," broke things like this:

  For example:
  * broken
  * microphone

That will now be recognized as a list again.  The heuristic is
now when two lines in a row start with the same type of list
marker then recognize that as a list even when the first item
doesn't appear to start its own paragraph.

Additionally if the second line is at the next indent level
then the two lines may have different kinds of list markers
and still be recognized.

All previously recognized lists are still recognized.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>