Markdown.pl: do not mishandle double list markers
This example:

  * a
      + 1. x
      + 2. y
  * c

Should format as one outer "ul" with the first item having
a second inner "ul".  There should not be any "ol" lists in
the formatted result at all.

Correct the code so that it does not think "+ 1. x" not only
starts a list but also includes a sublist.

Now it only starts a list where the first item just happens to
have content that closely resembles a list marker.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>