Markdown.pl: format fancy checkboxes
Unordered list items that begin with '[ ]' or '[x]' and
a space will be formatted to display a fancy checkbox
(and checkmark for "x") when the fancy style sheet is
included in the output.

Without the fancy stylesheet everything will still look
fine, but the fancy stuff won't be there.

A new "--show-stylesheet" option is added to show the
style sheet at the beginning of the output.  When combined
with no arguments and redirecting standard input to
/dev/null it can be used to show just the style sheet.

And, since we're adding a stylesheet, add an item to
make the "code" blocks look a bit nicer too.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>