Markdown: allow trailing '[]' to be omitted
When using the "implicit link name" shortcut (i.e. the
link name is the same as the link text), the trailing '[]'
is unsightly.

Allow the trailing '[]' to be omitted when the omission is
unambiguous.  In other words, if there is no preceding or
following pair of square brackets the trailing '[]' can safely
be omitted.

For example:

  See any of [link 1][] [link 2][] [lnik 3][].

The trailing '[]' MUST NOT be omitted in this case because the
result:

  See any of [link 1] [link 2] [link 3]

would be misinterpreted.  But, if they're separated with commas
or words instead like so:

  See any of [link 1], [link 2] or [link 3].

then they cannot be misinterpreted and the trailing '[]' can be
safely omitted making for a much nicer looking document.

To go with this change the basics.md and syntax.md documents
have been modified to take advantage of these new semantics.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>