Markdown.pl: do not mishandle double list markers redux
In b7f3fc1c (Markdown.pl: do not mishandle double list markers,
2017-01-09, markdown_1.1.0), an attempt was made to avoid
having something like "* 1. item" be misinterpreted as an
unordered list containing an ordered sublist.

However, the change made to fix the problem only introduced
another problem where lists were not always being recognized
when they should be.

Fix the fix (it did have a bit of a kludgely side to it) so
it works properly and is less kludgely.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>