Markdown.pl: normalize link white space
White space in '#', '##', '###', '####', '#####' and '######'
headers is already being normalized (i.e. leading and trailing
whitespace is stripped off).

Do the same thing for '=', '-', and '~' headers and, in addition,
do that for link ids, title text and alt text and for them also
replace internal runs of whitespace with a single space.

This makes the output nicer and more consistent and avoids
subtle bugs due to accidental inclusion of an extra space.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>