Markdown.pl: usually (i), (v) and (x) are roman
Only treat (i), (v) and (x) as alpha if the previous
list marker was lower alpha (or upper alpha in the
case of (I), (V) and (X)).

Previously they were treated as alpha if the first
marker in the list was alpha, but if list marker
types were changed mid-list that could lead to
unexpected behavior.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>