Markdown.pl: do not overlook sibling list items
When support for additional list markers was added in
51f3d63833 (Markdown.pl: support more list markers, 2017-01-10, 1.1.0),
a bug was inadvertently introduced that could cause adjacent sibling
list items to only recognize the first as a list item and as a side-effect
prevent markup from being recognized in the second.

The problem occurred when the matching pattern was split to run in
progressive matching mode and resulted in the sibling list items match
not always being matched by the progressive list item pattern (extra
possible \n's were preventing a match).

Fix this by adding a '+' in the correct location in the progressive pattern.

The side-effect was caused because any "leftover" (of which there shouldn't
be any) was not being processed for markup.

As a precaution, run any leftover through the block gamut markup processor
just in case even though there should never be any leftover.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>