Markdown.pl: keep valid tag list alphabetized
While trying to keep all the various table-related tags together
is admirable, it makes it hard to be sure the tag is in the list
or not (an also looks bad compared to the other tags).

Therefore put the table-related tags into alphabetical order
just like the rest of them.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>