Markdown.pl: disallow <dir> and <menu> without --deprecated
When dealing with program arguments "<dir>" is highly problematic.

Both "<dir>" and "<menu>" have long been deprecated and there are
other tags readily available for making similar lists that are not
deprecated (and do not require use of style sheets either).

Therefore treat "<dir>" and "<menu>" as literal text unless the
new "--deprecated" option is used.

Other "deprecated" tags continue to be recognized and passed through
as they generally do not have non-deprecated equivalents that do not
also require use of style attributes or style sheets in some fashion.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>