Markdown.pl: apply -i and -r options to a and img tags
When a and img tags are generated using the normal Markdown syntax
any prefixes specified with the -i and -r options are inserted as
appropriate.

Extend this processing to explicit a and img tags as well.

This makes sense because they should be handled the same way the
Markdown syntax generated tags are for consistency.

It's still possible to "escape" from the prefixes by using an
explicit scheme+host+port or the commonly supported (but not a
standard) //+host+port mechanism.

And it only matters if prefixes have been set with the -i and/or
-r options (the default is no prefixes) anyway.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>