Markdown: enhance link handling
Allow links of the form [...](#...) to find themselves on the
page the same way links of the form [...] can.

Be flexible accepting either '-' or '_' in place of spaces
in the heading name since fragment names may not contain spaces.

Refactor the code that manufactures "img" and "a" tags to both
simplify the code and make sure that all href, src, alt and title
attributes are fully and properly "escaped".

In addition, if the "title" for an image ends with something that
looks like "(512x342)", "(?x342)" or "(512x?)" then strip that out
of the title and set the appropriate width and height attributes
on the manufactured "img" tag.  For example something like this:

    ![Nice pic](pic.jpg "Nice (500x300)")

or this:

    ![Nice pic][1]
    [1]: <pic.jpg> "Nice (500x300)"

now produces this:

    <img src="pic.jpg" alt="Nice pic" width="500" height="300" title="Nice" />

Update the syntax doc to mention these additions.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>