Markdown.pl: allow documented single-quote ref titles
The documentation claims that all of these work:

    [1]: url1.txt "title 1"
    [2]: url2.txt 'title 2'
    [3]: url3.txt (title 3)

However, the single-quote version was not being accepted.

Update the regular expression to grok the single-quote
variant and require the correct matching closing quote
in order to match.  This corrects the invalid acceptance
of mismatched quote titles such as "title) and (title".

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>