Markdown.pl: finesse backslash handling in tables
Prevent loss of backslash-escaping in tables by avoiding an effective
de-backslash operation that was incorrect.  Now `\\\\` always ends up
producing two backslashes whether it's in a table or not.

Adjust the regexps so that the `|` in `\|` does not get mistaken for
the final `|` of a table row when that table row has omitted the
final trailing `|`.

Also silently handle a lone `\` at the end of a table row by pretending
it was doubled rather than breaking the table.

While in that area of the code, remove a line that computed a value
which was never used.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>