Markdown.pl: perform tag sanitation and take out the trash
Unless the new `--no-sanitize` option is given, then inspect each
raw tag encountered in the input and "sanitize" it before outputting
it.  The new `--sanitize` option is now the default.

As before, any tags not on the "approved" list have their "<" escaped
and become ordinary text in the output.  With this commit, two new
tags are now on the "approved" list:  `<map>` and `<area>`.

Each "approved" tag encountered has all of its attributes inspected
and any that are not on the "approved" list for that tag are expunged.
Any others that aren't in canonical form are corrected.

Empty tags are normalized as well (e.g. "<p/>" becomes "<p></p>"
but "<img ...>" becomes "<img ... />" etc.).

With this change, Markdown.pl output should be reasonably "safe"
for use as no code tags or attributes are permitted.

While client-side image maps are now allowed, server-side are not
(the "ismap" attribute will be silently expunged).  Client-side
image maps should always be preferred to server-side anyway, that's
why they were created in the first place.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>