Markdown.pl: correct tab expansion for shifted ``` blocks
A "backticks-delimited code block" must be at the left margin,
however, the entire block can be shifted 1, 2 or 3 characters to
the right using spaces.  These are all the same:

```
line
```

 ```
 line
 ```

  ```
  line
  ```

   ```
   line
   ```

When support for the slightly shifted code blocks was originally
added, the effect on tab expansion was overlooked.

The shifting in of the code block serves to make the raw markdown
source perhaps a tiny bit more readable, but the shifting in is NOT
a logical part of the lines in the code block themselves.

Therefore, remove any "shift in" spaces before expanding tabs within
the code block so that the result ends up looking exactly the same
no matter whether the code block has been shifted in at all or not.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>