Markdown.pl: --validate-xml-internal by default
Enhance the sanitation process slightly in order to perform simple
tag missing/mismatch validation.  Almost all the work needed was
already being performed with the exception of keeping a tag stack.

Keep an active tag stack when `--validate-xml-internal` is active
and use it to find mismatched and/or missing open/closing tags.

Enable this by default unless `--no-sanitize` has been given (the
sanitize machinery does the validation and is required) or
`--validate-xml` or `--no-validate-xml` has been given explicitly.

Unlike the more comprehensive `--validate-xml`, this option operates
very quickly and does not require any additional XML modules to be
present.  It's also compatible with `--html4tags`.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>