Markdown.pl: use g_nested_brackets for images too
The recursive regex `$g_nested_brackets` (see the source) correctly
matches properly balanced nested `[`...`]` text.

The normal anchor parsing already makes use of it.

Use it for the image parsing too so that this:

    [![Alt](img.png)][link]
    [link]:example.txt

Does not become the very broken:

    <a href="img.png"><img src="example.txt" alt="Alt</a>" />

But instead becomes the correct:

    <a href="example.txt"><img src="img.png" alt="Alt" /></a>

This problem has been present all the way back to Markdown.pl version
1.0.1 (2004-12-14) and likely earlier too.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>