Markdown.pl: fun with fragments
Add new `--base` option that allows a prefix to be specified to be
added to all bare fragment-only URL links.

Use of this option may be required in order for intra-document
fragment links to function properly within a document that makes
use of the `<base>` tag.

Make sure explicitly specified fragment-only URLs (i.e. given in
verbatim `<a>` tags) get hooked up to the proper destination if
possible.

They obviously are trying to refer to something in the same document
so make sure they get the same treatment to hook them up.

Do the same for fragment-only links inside wiki `[[`...`]]` links.

And for both of these (explicit `<a>` tags and `[[`...`]]` links)
make sure the new bare fragment-only URL prefix gets added if given.

While in there, adjust whitespace to match coding convention for
this file where needed in the sections that have been changed.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>