Markdown.pl: enhance --wiki "s" option operation
The "s" option of the --wiki format strips the final extension
before applying the template.

Enhance the "s" option to optionally take a list of extensions
and to only strip the extension if it's one from the list.

Provide a "shortcut" extension that represents all known markdown
extensions.

Change the default --wiki format to now be "%{s(:md)}.html" instead
of the previous default which means it will no longer strip arbitrary
extensions, but only known markdown ones.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>