Markdown.pl: support new --absroot=prefix option
Any absolute path URLs (but not // ones) have the prefix prepended.
If that makes the resulting URL a fully absolute URL it will not
be processed by any --htmlroot and/or --imageroot options.

With this option, site-relative absolute path URLs can be re-written
so that the site is made explicit in order to support viewing on
a different site.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>