Markdown.pl: recognize U+00D7 for "x"
When parsing a "checkbox" item or image dimensions, recognize
a U+00D7 Multiplication Sign character as equivalent to an "x".

The real "x" is preferred (and still recognized along with "X"),
but in the case where a U+00D7 (×) ends up in there, just go
with it and recognize it as the intent remains clear.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>