Markdown.pl: minor empty tag improvments
Even though block tags such as "<p/>" should not appear in
valid XHTML documents, the internal validator (which is
enabled by default) will properly expand "<p/>" to "<p></p>".

However, the block formatting code fails to notice such
an empty tag block leading to it being wrapped in a spurious
"<p>...</p>" pair before it's expanded by the validation code.

Attempt to recognize some of these valid-for-xml-but-not-xhtml
blocks earlier to produce better output.

This is not a perfect fix, but it's an improvement.

It's really an odd edge case anyway that's unlikely to be
encountered very often.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>