Markdown.pl: detail ability to "use" Markdown module
While the default mode of Markdown.pl remains that of a command
line utility, it's fairly simple to "use Markdown" and call the
functions directly.

Explain this usage in the help and make sure all of the auxiliary
functions that might be used for this appear in @EXPORT_OK.

Include an example that simulates `Markdown.pl --stub --wiki`.

Add a symbolic link from Markdown.pm to Markdown.pl to go
along with the new example.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>