Markdown.pl: improve standalone XML comment stripping
When stripping XML comments, if any XML comments are recognized as
a standalone block, strip that entire block when forming paragraphs
the final time.

This provides a much cleaner output as it results in many
superfluous blank lines being suppressed that the XML parser
would not otherwise remove when it strips out XML comments.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>