Markdown: support floating images
With a minor enhancement to the support for specifying image
dimensions, images can now be "float"ed to the left or right
or even centered in their own block.

Add the ability to generate a <br clear="all" /> with 3 or
more spaces on the end of a line rather than a plain <br />
with only 2.

Document these additions as well.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>