Markdown.pl: avoid superfluous leading <p></p>
Since each "paragraph" is wrapped between a "<p>" and "</p>"
this input:

    <p>hi

    <p>bye

has been producing this output:

    <p></p><p>hi</p>

    <p></p><p>bye</p>

Correct this so that if the leading "<p>" of the paragraph wrapper
is immediately auto-closed then it's simply discarded rather than
creating a bogus "<p></p>" section.

With this change the previous input now produces this output:

    <p>hi</p>

    <p>bye</p>

The bogus leading "<p></p>" sections have been omitted and the
output looks much nicer.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>