Markdown.pl: dd, dt and li do not autoclose containing table
While <dd>, <dt> and <li> all have "optional" closing tags, they
can all be contained within a table.

And as such must not close the tags that define the content of
the table itself.

Customize the tagacl list for these three to exclude the tags
that may contain table content to prevent their premature closing.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>