Markdown.pl: sanitize all "&" issues
When sanitize is active (--sanitize, the default), make sure all
"&" issues are checked.  This includes things like bare "&" that
should be "&amp;" but aren't.  And it includes single/double
quote characters inside attribute values that should be encoded
and are not.

Since the internal validator requires the sanitize mode to be
active, this now makes sure that the internal validation mode
cannot pass through any invalid entity references to the output.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>