Markdown.pl: document and catch more meaningless tags
There are a few tags (e.g. `a`, `area`, `img`, `map`) that
require at least one attribute to be present in order to be
meaningful.

When these tags occur without any attributes they are treated
as non-tags and the leading `<` is escaped to `&lt;`.

This can only happen when sanitize mode is active.

Although already partially implemented, it was not documented
in the help.

Add discussion of this to the help and make the implementation
more robust to catch more of these tags.

This is not intended to be a perversely pedantic change, but
rather to allow such meaningless tags to be used as plain text
without the need for escaping.  For example the text:

    The <a><c><e> process ...

Can be used exactly as-is and all of the `<`s will automatically
be escaped to `&lt;` since none of them specify meaningful tags.

Of course, using the `--no-sanitize` option will disable this
behavior.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>