Markdown.pl: provide urlfunc hook and helpers
Provide a new urlfunc hook that can inspect/change all urls that
are in "a" "href" attributes and "img" "src" attributes.

Make the new SplitURL and unescapeXML routines exportable (@EXPORT_OK)
and rename the old escape function to be escapeXML and make
it exportable (@EXPORT_OK) too.

Add some nice comments to each of the newly exportable functions.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>