Markdown.pl: isolate the archaic tab default
If running as a plug in for either of the two original systems
that this was designed to "plug in" to, continue to use the
archaic, non-standard default for expansion width of physical
tabs.  This setting does not affect the "indent level" width.

Otherwise, force the physical tab width expansion to default
to the expected and standard value.

This has been the behavior for some time already, except that
when "use"ing Markdown.pm and calling the API directly this
was being bypassed in favor of the old, archaic default.

With this change, the old, archaic default becomes isolated
to those two originally supported systems.

The setting can still, of course, be changed by using an option
to whatever is desired.  The default though will now be more
sane for more clients.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>