Markdown.pl: strip markup out of implicit anchors
Each H1, H2, ... H6 generated courtesy of markdown markup has an
implicit anchor assigned based on the content of the element.

For example:

   # This is an _H1_ header

Strip any inline markup (in this case the '_'s) out before creating
the implicit anchor.  With this change, the text used to generate
the anchor for the above is just "This is an H1 header".

There are a couple of additional places where text that might have
inline markup gets turned into an identifier (implicit reference
links such as [thing][] or [thing] and wiki links without an
explicit link destination such as [[thing]]).  Perform the same
tag stripping for them too before trying to find the destination.

Many links that should have connected previously now do.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>